[pcj/logarithm] Revert ln(0) raising undef.
Only return nan in approximation for now. This is an intermediary step because this raise was blocking a few upcoming simplifications.

We should indeed project ln into a special tree on projection, that will be replaced with Ln and add a dependency as well.

But we need to handle and simplify exp(ln(0)) when it comes from power for instance.