refactor(approximation): use signaling nan to indicate nonreal
Because otherwise when we useed s_context->m_isNonReal, we needed to
return a non NAN value because when a value is NAN (really undef), we
don't want to return nonreal.

This created a lot of problem. Now, with signaling nan, the std::complex
returned for nonreal and undef in NAN.